+++ groups/bsl/bslstl/bslstl_stack.t.cpp
@@ -1996,7 +1996,7 @@ void TestDriver<CONTAINER>::testCase7()
             const size_t      LENGTH = (int) strlen(SPEC);
 
             if (verbose) {
-                printf("\nFor an object of length %d:\n", LENGTH);
+                printf("\nFor an object of length %d:\n", (int)LENGTH);
                 P(SPEC);
             }
 
@@ -2536,7 +2536,7 @@ void TestDriver<CONTAINER>::testCase3()
             Obj mX(&oa);
 
             if ((int)LENGTH != oldLen) {
-                if (verbose) printf("\tof length %d:\n", LENGTH);
+                if (verbose) printf("\tof length %d:\n", (int)LENGTH);
                  ASSERTV(LINE, oldLen <= (int)LENGTH);  // non-decreasing
                 oldLen = LENGTH;
             }
